+++ cmd/service/main.go
@@ -6,16 +6,15 @@ import (
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 
 	"github.com/go-chi/chi/v5"
-
 	"github.com/prometheus/client_golang/prometheus"
 	"golang.org/x/sync/errgroup"
 
 	server "github.com/lidofinance/finding-forwarder/internal/app/http_server"
 	"github.com/lidofinance/finding-forwarder/internal/connectors/logger"
 	"github.com/lidofinance/finding-forwarder/internal/connectors/metrics"
+	redisConnector "github.com/lidofinance/finding-forwarder/internal/connectors/redis"
 	"github.com/lidofinance/finding-forwarder/internal/env"
 )
 
@@ -35,14 +34,16 @@ func main() {
 		return
 	}
 
+	red := redisConnector.New(&cfg.AppConfig)
+
 	log.Info(fmt.Sprintf(`started %s application`, cfg.AppConfig.Name))
 
 	r := chi.NewRouter()
-	metrics := metrics.New(prometheus.NewRegistry(), cfg.AppConfig.Name, cfg.AppConfig.Env)
+	promStore := metrics.New(prometheus.NewRegistry(), cfg.AppConfig.Name, cfg.AppConfig.Env)
 
 	usecase := server.Usecase(&cfg.AppConfig)
 
-	app := server.New(log, metrics, usecase)
+	app := server.New(log, promStore, usecase, red)
 
 	app.Metrics.BuildInfo.Inc()
 	app.RegisterRoutes(r)
@@ -50,7 +51,7 @@ func main() {
 	g, gCtx := errgroup.WithContext(ctx)
 
 	app.RunHTTPServer(gCtx, g, cfg.AppConfig.Port, r)
-	// someDaemon(gCtx, g)
+	// someDaemon(gCtx, g, red)
 
 	if err := g.Wait(); err != nil {
 		log.Error(err)
@@ -59,15 +60,38 @@ func main() {
 	fmt.Println(`Main done`)
 }
 
-func someDaemon(gCtx context.Context, g *errgroup.Group) {
+/*func someDaemon(gCtx context.Context, g *errgroup.Group, redisClient *redis.Client) {
 	g.Go(func() error {
 		for {
 			select {
 			case <-time.After(1 * time.Second):
-				fmt.Println(2)
+
+				// Get the message
+				body, err := redisClient.LIndex(gCtx, topic, - 1).Bytes()
+				iferr ! =nil && !errors.Is(err, redis.Nil) {
+				return err
+				}
+				// No message, sleep for a while
+				if errors.Is(err, redis.Nil) {
+					time.Sleep(time.Second)
+					continue
+				}
+				// Process the message
+				err = h(&Msg{
+					Topic: topic,
+					Body:  body,
+				})
+				iferr ! =nil {
+				continue
+			}
+				// If the processing succeeds, delete the message
+				iferr := q.client.RPop(ctx, topic).Err(); err ! =nil {
+				return err
+			}
+
 			case <-gCtx.Done():
 				return nil
 			}
 		}
 	})
-}
+}*/