+++ internal/app/http_server/routes.go
@@ -5,8 +5,6 @@ import (
 	"github.com/go-chi/chi/v5/middleware"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 
-	_ "github.com/prometheus/client_golang/prometheus"
-
 	"github.com/lidofinance/finding-forwarder/internal/http/handlers/alerts"
 	"github.com/lidofinance/finding-forwarder/internal/http/handlers/health"
 )
@@ -20,7 +18,7 @@ func (a *App) RegisterRoutes(r chi.Router) {
 	r.Get("/health", health.New().Handler)
 	r.Get("/metrics", promhttp.Handler().ServeHTTP)
 
-	alertsH := alerts.New(a.Logger)
+	alertsH := alerts.New(a.Logger, a.RedisClient, "test")
 
 	r.Post("/alerts", alertsH.Handler)
 }