+++ internal/app/http_server/usecase.go
@@ -1,17 +1,17 @@
 package server
 
 import (
+	"net/http"
+
 	"github.com/lidofinance/finding-forwarder/internal/env"
 	"github.com/lidofinance/finding-forwarder/internal/pkg/telegram"
 	usersUsecase "github.com/lidofinance/finding-forwarder/internal/pkg/telegram/usecase"
-	"net/http"
 )
 
 type usecase struct {
 	User telegram.Usecase
 }
 
-// nolint
 func Usecase(cfg *env.AppConfig) *usecase {
 	return &usecase{
 		User: usersUsecase.New(cfg.TelegramBotToken, http.Client{}),