+++ internal/connectors/metrics/prometheus.go
@@ -15,7 +15,7 @@ type Store struct {
 	BuildInfo  prometheus.Counter
 }
 
-func New(promRegistry *prometheus.Registry, appName string, env string) *Store {
+func New(promRegistry *prometheus.Registry, appName, env string) *Store {
 	store := &Store{
 		Prometheus: promRegistry,
 		BuildInfo: promauto.NewCounter(prometheus.CounterOpts{