+++ internal/env/env.go
@@ -11,14 +11,21 @@ type Config struct {
 }
 
 type AppConfig struct {
-	Name             string
-	Env              string
-	URL              string
-	Port             uint
-	LogFormat        string
-	LogLevel         string
-	TelegramBotToken string
-	TelegramChatID   string
+	Name              string
+	Env               string
+	URL               string
+	Port              uint
+	LogFormat         string
+	LogLevel          string
+	TelegramBotToken  string
+	TelegramChatID    string
+	OpsGeniaAPIKey    string
+	DiscordWebHookURL string
+
+	RedisURL       string
+	RedisPassword  string
+	RedisDB        int
+	RedisQueueName string
 }
 
 var (
@@ -33,7 +40,7 @@ func Read(configPath string) (*Config, error) {
 	onceDefaultClient.Do(func() {
 		viper.SetConfigType("env")
 
-		if len(configPath) != 0 {
+		if configPath != "" {
 			viper.SetConfigFile(configPath)
 		} else {
 			viper.AddConfigPath(".")
@@ -41,23 +48,27 @@ func Read(configPath string) (*Config, error) {
 		}
 
 		viper.AutomaticEnv()
-		if viperErr := viper.ReadInConfig(); err != nil {
-			if _, ok := viperErr.(viper.ConfigFileNotFoundError); !ok {
-				err = viperErr
-				return
-			}
+		if viperErr := viper.ReadInConfig(); viperErr != nil {
+			err = viperErr
+			return
 		}
 
 		cfg = Config{
 			AppConfig: AppConfig{
-				Name:             viper.GetString("APP_NAME"),
-				Env:              viper.GetString("ENV"),
-				URL:              viper.GetString("app.url"),
-				Port:             viper.GetUint("PORT"),
-				LogFormat:        viper.GetString("LOG_FORMAT"),
-				LogLevel:         viper.GetString("LOG_LEVEL"),
-				TelegramBotToken: viper.GetString("TELEGRAM_BOT_TOKEN"),
-				TelegramChatID:   viper.GetString("TELEGRAM_CHAT_ID"),
+				Name:              viper.GetString("APP_NAME"),
+				Env:               viper.GetString("ENV"),
+				Port:              viper.GetUint("PORT"),
+				LogFormat:         viper.GetString("LOG_FORMAT"),
+				LogLevel:          viper.GetString("LOG_LEVEL"),
+				TelegramBotToken:  viper.GetString("TELEGRAM_BOT_TOKEN"),
+				TelegramChatID:    viper.GetString("TELEGRAM_CHAT_ID"),
+				OpsGeniaAPIKey:    viper.GetString("OPSGENIE_API_KEY"),
+				DiscordWebHookURL: viper.GetString("DISCORD_WEBHOOK_URL"),
+
+				RedisURL:       viper.GetString("REDIS_URL"),
+				RedisPassword:  viper.GetString("REDIS_PASSWORD"),
+				RedisDB:        viper.GetInt("REDIS_DB"),
+				RedisQueueName: viper.GetString("REDIS_QUEUE_NAME"),
 			},
 		}
 	})