+++ internal/pkg/discord/usecase.go
@@ -1,8 +1,8 @@
-package telegram
+package discord
 
 import "context"
 
 //go:generate ./../../../bin/mockery --name Usecase
 type Usecase interface {
-	SendMessage(ctx context.Context, chatID, message string) error
+	SendMessage(ctx context.Context, message string) error
 }