+++ internal/pkg/opsgenia/usecase.go
@@ -1,8 +1,8 @@
-package telegram
+package opsgenia
 
 import "context"
 
 //go:generate ./../../../bin/mockery --name Usecase
 type Usecase interface {
-	SendMessage(ctx context.Context, chatID, message string) error
+	SendMessage(ctx context.Context, message, description, priority string) error
 }