+++ Changelog.md
@@ -1,5 +1,9 @@
 ## 31.05.2024
 1. Added method for sending messages into telegram chat
+2. Added method for sending messages into discord chat
+3. Added method for sending messages into opsGenia chat
+4. Set up linter rules
+5. Preparation for redis-queue task
 
 ## 30.05.2024
 1. Forked from go-template